+++ src/app/page.tsx
@@ -1,10 +1,8 @@
-import { AreaChartTypeExample } from "@/components/AreaChartTest";
+"use client"
+import { LoginPage } from '@/features/Login/Login/LoginPage';
 
 export default function Home() {
   return (
-    <div>
-      <h1>Home</h1>
-      <AreaChartTypeExample />
-    </div>
+    <LoginPage />
   );
 }