+++ CHANGELOG.md
@@ -1,3 +1,45 @@
+## v0.8.0 (2025-06-10)
+
+### Pull Requests
+[#11](https://github.com/RafaelMoro/next-budget-master/pull/11) | Add reset password page
+
+
+## v0.7.0 (2025-06-09)
+
+### Pull Requests
+[#10](https://github.com/RafaelMoro/next-budget-master/pull/10) | Add forgot password
+
+
+## v0.6.1 (2025-06-08)
+
+### Pull Requests
+[#9](https://github.com/RafaelMoro/next-budget-master/pull/9) | Add react hook login
+
+
+## v0.6.0 (2025-06-08)
+
+### Pull Requests
+[#8](https://github.com/RafaelMoro/next-budget-master/pull/8) | Add Register page
+
+
+## v0.5.0 (2025-06-07)
+
+### Pull Requests
+[#7](https://github.com/RafaelMoro/next-budget-master/pull/7) | Save cookie information from login
+
+
+## v0.4.0 (2025-06-06)
+
+### Pull Requests
+[#6](https://github.com/RafaelMoro/next-budget-master/pull/6) | Add Login screen
+
+
+## v0.3.0 (2025-06-05)
+
+### Pull Requests
+[#5](https://github.com/RafaelMoro/next-budget-master/pull/5) | Add flowbite and toggle theme mode
+
+
 ## v0.2.2 (2025-06-05)
 
 ### Pull Requests